requirements.txt
absl-py==2.1.0
APScheduler==3.10.4
astunparse==1.6.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorhash==2.0.0
configparser==7.0.0
Flask==3.0.3
Flask-MonitoringDashboard==3.3.2
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
plotly==5.22.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==6.0.0
pyctuator==1.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
rich==13.7.1
scipy==1.13.1
six==1.16.0
SQLAlchemy==2.0.31
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
